bugfix: error during configure leads to NULL ptr
When TLS_Socket sets an error with set_eror(), it should not try to
finish closing the socket if the state is TLS_SOCKET_STATE_NONE.

Explanation:

In the TLS_SOCKET_STATE_NONE state, the socket is not configured
correctly. Some or all of the mbedtls state may be undefined or NULL at
this point, so calling finish_close() (which calls mbedtls_ssl_session_reset()) does not make sense.